requirements.txt
pandas==1.4.0
beautifulsoup4==4.9.3
requests==2.27.1
Django==2.2
django-background-tasks==1.2.5
